backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
Pillow==11.2.1
